+++ src/collision/OverlapCallback.cpp
@@ -82,7 +82,7 @@ OverlapCallback::CallbackData::CallbackData(Array<ContactPair>& contactPairs, Ar
     }
     // Filter the lost contact pairs to only keep the overlap/trigger events (not the contact events)
     const uint64 nbLostContactPairs = mLostContactPairs.size();
-    for (uint i=0; i < nbLostContactPairs; i++) {
+    for (uint64 i=0; i < nbLostContactPairs; i++) {
 
         // If the contact pair contains contacts (and is therefore not an overlap/trigger event)
         if (!onlyReportTriggers || mLostContactPairs[i].isTrigger) {
